fix: absent and acl mapping

+++ src/routes/MasterData.ts
@@ -16,32 +16,42 @@ import * as MahasiswaValidation from "$validations/master-data/mahasiswa";
 const MasterDataRoutes = new Hono();
 
 // ======= Shift =======
-MasterDataRoutes.get("/shift", AuthMiddleware.checkJwt, ShiftController.getAll);
+MasterDataRoutes.get(
+    "/shift",
+    AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "VIEW"),
+    ShiftController.getAll
+);
 MasterDataRoutes.get(
     "/shift/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "VIEW"),
     ShiftController.getById
 );
 MasterDataRoutes.post(
     "/shift",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "CREATE"),
     ShiftValidation.validateShift,
     ShiftController.create
 );
 MasterDataRoutes.put(
     "/shift/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "UPDATE"),
     ShiftValidation.validateShift,
     ShiftController.update
 );
 MasterDataRoutes.delete(
     "/shift",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "DELETE"),
     ShiftController.deleteByIds
 );
 MasterDataRoutes.put(
     "/shift/:id/activate",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("SHIFT", "DELETE"),
     ShiftController.activate
 );
 // =====================
@@ -50,28 +60,33 @@ MasterDataRoutes.put(
 MasterDataRoutes.get(
     "/ruangan",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "VIEW"),
     RuanganController.getAll
 );
 MasterDataRoutes.post(
     "/ruangan",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "CREATE"),
     RuanganValidation.validateRuangan,
     RuanganController.create
 );
 MasterDataRoutes.get(
     "/ruangan/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "VIEW"),
     RuanganController.getById
 );
 MasterDataRoutes.put(
     "/ruangan/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "UPDATE"),
     RuanganValidation.validateRuangan,
     RuanganController.update
 );
 MasterDataRoutes.put(
     "/ruangan/:id/assign-kepala-lab",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "UPDATE"),
     RuanganValidation.validateAssignKepalaLab,
     RuanganController.assignKepalaRuangan
 );
@@ -79,6 +94,7 @@ MasterDataRoutes.put(
 MasterDataRoutes.put(
     "/ruangan/:id/activate",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("RUANGAN", "DELETE"),
     RuanganController.activate
 );
 // =====================
@@ -87,33 +103,39 @@ MasterDataRoutes.put(
 MasterDataRoutes.get(
     "/mata-kuliah",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "VIEW"),
     MataKuliahController.getAll
 );
 MasterDataRoutes.get(
     "/mata-kuliah/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "VIEW"),
     MataKuliahController.getById
 );
 MasterDataRoutes.post(
     "/mata-kuliah",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "CREATE"),
     MataKuliahValidation.validateMatakuliah,
     MataKuliahController.create
 );
 MasterDataRoutes.put(
     "/mata-kuliah/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "UPDATE"),
     MataKuliahValidation.validateMatakuliah,
     MataKuliahController.update
 );
 MasterDataRoutes.delete(
     "/mata-kuliah",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "DELETE"),
     MataKuliahController.deleteByIds
 );
 MasterDataRoutes.post(
     "/mata-kuliah/bulk-upload",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MATA_KULIAH", "BULK"),
     MataKuliahController.bulkUpload
 );
 // =====================
@@ -122,64 +144,80 @@ MasterDataRoutes.post(
 MasterDataRoutes.get(
     "/mahasiswa",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "VIEW"),
     MahasiswaController.getAll
 );
 MasterDataRoutes.get(
     "/mahasiswa/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "VIEW"),
     MahasiswaController.getById
 );
 MasterDataRoutes.post(
     "/mahasiswa",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "CREATE"),
     MahasiswaValidation.validateMahasiswa,
     MahasiswaController.create
 );
 MasterDataRoutes.put(
     "/mahasiswa/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "UPDATE"),
     MahasiswaValidation.validateMahasiswa,
     MahasiswaController.update
 );
 MasterDataRoutes.delete(
     "/mahasiswa",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "DELETE"),
     MahasiswaController.deleteByIds
 );
 MasterDataRoutes.post(
     "/mahasiswa/bulk-upload",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("MAHASISWA", "BULK"),
     MahasiswaController.bulkUpload
 );
 // =====================
 
 // ======= Dosen =======
-MasterDataRoutes.get("/dosen", AuthMiddleware.checkJwt, DosenController.getAll);
+MasterDataRoutes.get(
+    "/dosen",
+    AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "VIEW"),
+    DosenController.getAll
+);
 MasterDataRoutes.get(
     "/dosen/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "VIEW"),
     DosenController.getById
 );
 MasterDataRoutes.post(
     "/dosen",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "CREATE"),
     DosenValidation.validateDosen,
     DosenController.create
 );
 MasterDataRoutes.put(
     "/dosen/:id",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "UPDATE"),
     DosenValidation.validateDosenUpdate,
     DosenController.update
 );
 MasterDataRoutes.delete(
     "/dosen",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "DELETE"),
     DosenController.deleteByIds
 );
 MasterDataRoutes.post(
     "/dosen/bulk-upload",
     AuthMiddleware.checkJwt,
+    AuthMiddleware.checkAccess("DOSEN", "BULK"),
     DosenController.bulkUpload
 );
 // =====================
